refactor(server): extract citation-source research → src/server/citations.js
Pure move, zero behavior change. Pulls findCitationSources (+ its private
LOW_QUALITY_CITATION_DOMAINS blocklist) out of server.js into
src/server/citations.js (~90 lines).

Why: continues the server.js dismemberment. findCitationSources is a
cohesive unit — Perplexity Sonar source-finding + domain-quality /
self-domain filtering. Its only non-global dep is the shared pool (brand-
domain lookup), imported from db.js. No other local helpers.

What:
- New src/server/citations.js exports findCitationSources (consumed by the
  compliance find-sources + verify-and-rewrite routes); imports pool from
  db.js. LOW_QUALITY_CITATION_DOMAINS stays module-private (used only
  inside findCitationSources).
- server.js imports findCitationSources; inline defs -> breadcrumb.

Test plan:
- node --check server.js + citations.js -> OK
- npm run lint (no-undef gate) -> clean (confirms findCitationSources
  resolves and no orphaned LOW_QUALITY_CITATION_DOMAINS reference remains)
- route-inventory guard -> PASS (snapshot unchanged)
- vitest -> 78/78 pass. No dedicated unit test: findCitationSources is
  Perplexity-fetch + DB-query bound (like uploadXMedia / scrape), exercised
  via the live find-sources path.

Rollback: revert; pure copy, restores inline defs exactly.
